pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "backend for telegram calendar"
authors = ["hillow <andrew.hillow@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.21.1"
python-telegram-bot = "^20.0"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
tortoise-orm = "^0.19.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
